Update certificate-connector-prerequisites.md
If installing the connector on a Windows server 2019 or older and if Microsoft Edge is not installed, we will experience a warning in the step Azure AD Sign to "Keep your account secure". The recommendation in the warning is to update the browser or install the latest version of Microsoft Edge. Installing it beforehand fix's this issue. Adjusting the Note to reflect this.