Update certificate-connector-install.md
If installing the connector on a Windows server 2019 or older and if Microsoft Edge is not installed, we will experience a warning in the Azure AD Sign step to "Keep your account secure". The recommendation in the warning is to update the browser or install the latest version of Microsoft Edge. Installing it beforehand fix's this issue. Adding a Note to warn administrators about this.