Update apps-win32-add.md
This document doesn't say about "global reevaluation schedule". Our customers often want to know the word and raise some support tickets.
Based on the below document about GRS, I added several phrases for clarification

https://supportability.visualstudio.com/Intune/_wiki/wikis/Intune/1990305/Windows-IME?anchor=grs---global-re-evaluation-schedule
//
The installation is repeated 3 times. However, there must be at least 5 minutes between the attempts. When all these 3 attempts fail then at the app is in the GRS (Global Re-evaluation Schedule). This means the IME waits 24h to try the next attempt to install the app. You can see this in the log with this line: