Update azure-virtual-desktop-multi-session.md
In a case we have with EON '2502280060005951' we found this limitation is expected as confirmed with Intune PG in this ICM: https://portal.microsofticm.com/imp/v5/incidents/details/695388974/summary. When the user is added to the include scope, the registry values remain for later excluded users if they have FSLogix profiles and the user isn't signed in during exclusion. The exclusion will apply every time in personal hostpool sessions with and or without FSLogix profile. The exclusion will apply every time if the user uses local profile and doesn't have FSLogix profile in personal and pooled. The exclusion will apply if the user has FSLogix profile in pooled hostpool and is logged on during applying the Intune exclusion.

This needs to be documented so that the customers accept the fact that it is expected.